feat: added basic simulation for the automaton
This is a prototype that I was just working on, I never thought of it as being the final thing, hence I never tracked its creation using Git. I think this is what the whole of this project has been; an experiment. But since its working fine, I'll try polishing the whole thing up.